# docs/swagger.yaml
definitions:
  requests.RefreshTokensRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  requests.SignInWithWalletRequest:
    properties:
      wallet_pub:
        type: string
    required:
    - wallet_pub
    type: object
  requests.VerifySignatureRequest:
    properties:
      signature:
        type: string
      wallet_pub:
        type: string
    required:
    - signature
    - wallet_pub
    type: object
  response.ErrorResponse:
    properties:
      error:
        type: string
    required:
    - error
    type: object
  response.RefreshTokensResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    required:
    - access_token
    - refresh_token
    type: object
  response.SignInWithWalletResponse:
    properties:
      nonce:
        type: string
    required:
    - nonce
    type: object
  response.VerifySignatureResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    required:
    - access_token
    - refresh_token
    type: object
info:
  contact: {}
paths:
  /api/auth/refresh-tokens:
    post:
      consumes:
      - application/json
      description: Обновление access и refresh токенов с использованием валидного
        refresh токена
      parameters:
      - description: RefreshTokens Request
        in: body
        name: refreshTokens
        required: true
        schema:
          $ref: '#/definitions/requests.RefreshTokensRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Новые access и refresh токены
          schema:
            $ref: '#/definitions/response.RefreshTokensResponse'
        "400":
          description: Неверный refresh токен или ошибка валидации
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Обновление Access и Refresh токенов
      tags:
      - auth
  /api/auth/sign-in:
    post:
      consumes:
      - application/json
      description: Авторизация пользователя с использованием его Ethereum кошелька
      parameters:
      - description: SignInWithWallet Request
        in: body
        name: signInWithWallet
        required: true
        schema:
          $ref: '#/definitions/requests.SignInWithWalletRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Сгенерированный uuid для проверки подписи (nonce)
          schema:
            $ref: '#/definitions/response.SignInWithWalletResponse'
        "400":
          description: Ошибка валидации или неправильный запрос
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Авторизация через кошелек
      tags:
      - auth
  /api/auth/verify-signature:
    post:
      consumes:
      - application/json
      description: Проверка подписи с использованием публичного ключа
      parameters:
      - description: VerifySignature Request
        in: body
        name: verifySignature
        required: true
        schema:
          $ref: '#/definitions/requests.VerifySignatureRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Подпись успешно проверена
          schema:
            $ref: '#/definitions/response.VerifySignatureResponse'
        "400":
          description: Неверная подпись или неправильный запрос
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Верификация подписи
      tags:
      - auth
swagger: "2.0"